// src/module/contact-info/contact.interface.ts
export interface InstitutionData {
  institutionNameEnglish: string;
  institutionAddressEnglish: string;
  institutionNameBanglaArabic: string;
  institutionAddressBanglaArabic: string;
  address: string;
  contactNumber: string;
  whatsApp: string;
  telegram: string;
  facebook: string;
  twitter: string;
  instagram: string;
  youtube: string;
  discord: string;
  messenger: string;
  linkedIn: string;
  submissionDate: string;
  logo: string;
}